[ethers-v5] refactor: code split the function methods
